First working version
This commit adds a simple Python script and a Dockerfile to create a
minimal container image around it. The script expects to run as part of
a Pod in Kubernetes with the stock CouchDB container. When it runs it
will query DNS looking for the SRV record associated with a headless
service called "couchdb". It will then iterate over the targets of that
SRV record and create well-formed membership documents for each one in
the local nodes DB.

The end result should be that all the Pods backing a StatefulSet
associated with the headless Service will be automatically joined up
into a cluster on startup.